refactor(logic&api): 로직 및 API 분리 * 결과 페이지 로직 분리 * 영화 API 상수화 Related to: #25
